Set brightness attribute to double type
It seems that Z-Wave lights have the brightness as a float value, which
causes the app to crash. This should be fixed on the component side,
but it doesn't hurt to handle it in the app as well. Fixes #19.

+++ app/src/main/java/io/homeassistant/android/api/results/Attributes.java
@@ -5,7 +5,7 @@ import com.afollestad.ason.AsonName;
 import java.util.List;
 
 public class Attributes {
-    public int brightness;
+    public double brightness;
     @AsonName(name = "entity_id")
     public String[] children = null;
     public String entity_picture;
